backend/requirements.txt
openai-whisper
torch
fastapi
uvicorn
python-multipart
qdrant-client
requests
pytest
pytest-cov
